Rename scheduler.{algorithm -> podschedulers}
