+++ contrib/mesos/pkg/scheduler/podschedulers/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package algorithm
+package podschedulers
 
 import (
 	"errors"
